requirements.txt
python-decouple
telethon
requests
beautifulsoup4